Add NetworkLogger.enableProxy
